requirements.txt
tensorflow-gpu >= 2.10.0
keras >= 2.10.0
numpy >= 1.24.0
h5py >= 3.7.0
scikit-learn >= 1.2.0
tqdm >= 4.66.0
mealpy >= 2.5.3